requirements.txt
streamlit
openai
elevenlabslib
youtube-transcript-api
uuid
python-dotenv
